+++ docs/source/release-notes.rst
@@ -1,6 +1,17 @@
 Release notes
 =============
 
+0.0.0.8
+-------
+
+Add possibility to plot parameters vs distance from the center of the molecule
+to the center of the bond (see 
+:ref:`mode <tb2j_plotter_mode>` and 
+:ref:`substrate_atoms <tb2j_plotter_substrate_atoms>`).
+
+Add argument to ``tb2j_plotter.py`` for title for the pictures 
+(see :ref:`title <tb2j_plotter_title>`).
+
 0.0.0.7
 -------
 
@@ -9,8 +20,8 @@ Add the :doc:`phonopy_plotter.py <phonopy_plotter>` script
 0.0.0.6
 -------
 
-Add arguments :ref:`--scale-data <tb2j_plotter_scale_data>` and 
-:ref:`--scale-atoms <tb2j_plotter_scale_atoms>` to the ``tb2j_plotter.py``.
+Add arguments :ref:`scale-data <tb2j_plotter_scale_data>` and 
+:ref:`scale-atoms <tb2j_plotter_scale_atoms>` to the ``tb2j_plotter.py``.
 
 0.0.0.5
 -------